move balanced lloyd to graph.  move supporting functions.
